requirements.txt
dwave-neal==0.5.9
dwave-system==1.12.0
torch==1.10.2
